Moving all wiring into a single file #176
